call stats in its own file
