+++ stream-video-android-core/src/main/kotlin/io/getstream/video/android/core/CallStats.kt
@@ -1,0 +1,149 @@
+package io.getstream.video.android.core
+
+import android.os.Build
+import io.getstream.log.taggedLogger
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import org.webrtc.CameraEnumerationAndroid
+import org.webrtc.RTCStats
+import org.webrtc.RTCStatsReport
+import stream.video.sfu.models.TrackType
+
+
+
+
+public class PeerConnectionStats() {
+    internal val _resolution: MutableStateFlow<String> = MutableStateFlow("")
+    val resolution: StateFlow<String> = _resolution
+
+    internal val _qualityDropReason: MutableStateFlow<String> = MutableStateFlow("")
+    val qualityDropReason: StateFlow<String> = _qualityDropReason
+
+    internal val _jitterInMs: MutableStateFlow<Int> = MutableStateFlow(0)
+    val jitterInMs: StateFlow<Int> = _jitterInMs
+
+    internal val _bitrateKbps: MutableStateFlow<Float> = MutableStateFlow(0F)
+    val bitrateKbps: StateFlow<Float> = _bitrateKbps
+}
+
+public data class LocalStats(
+    val resolution: CameraEnumerationAndroid.CaptureFormat?,
+    val availableResolutions: List<CameraEnumerationAndroid.CaptureFormat>?,
+    val maxResolution: CameraEnumerationAndroid.CaptureFormat?,
+    val sfu: String,
+    val os: String,
+    val sdkVersion: String,
+    val deviceModel: String,
+)
+
+
+public class CallStats(val call: Call) {
+    private val logger by taggedLogger("CallStats")
+
+    val publisher = PeerConnectionStats()
+    val subscriber = PeerConnectionStats()
+    val _local = MutableStateFlow<LocalStats?>(null)
+    val local : StateFlow<LocalStats?> = _local
+
+    fun updateFromRTCStats(stats: RTCStatsReport?, isPublisher: Boolean = true) {
+        if (stats == null) return
+        // also see https://github.com/GetStream/stream-video-js/blob/main/packages/client/src/stats/state-store-stats-reporter.ts
+
+        val skipTypes = listOf("codec", "certificate", "data-channel")
+        val trackToParticipant = call.session?.trackIdToParticipant?.value ?: emptyMap()
+        val displayingAt = call.session?.trackDimensions?.value ?: emptyMap()
+
+        val statGroups = mutableMapOf<String, MutableList<RTCStats>>()
+
+        for (entry in stats.statsMap) {
+            val stat = entry.value
+
+            val type = stat.type
+            if (type in skipTypes) continue
+
+            val statGroup = if (type == "inbound-rtp") {
+                "$type:${stat.members["kind"]}"
+            } else if (type == "track") {
+                "$type:${stat.members["kind"]}"
+            } else if (type == "outbound-rtp") {
+                val rid = stat.members["rid"] ?: "missing"
+                "$type:${stat.members["kind"]}:$rid"
+            } else {
+                type
+            }
+
+            if (statGroup != null) {
+                if (statGroup !in statGroups) {
+                    statGroups[statGroup] = mutableListOf()
+                }
+                statGroups[statGroup]?.add(stat)
+            }
+
+            statGroups["outbound-rtp:video:f"]?.firstOrNull()?.let {
+                val qualityLimit = it.members["qualityLimitationReason"] as? String
+                val width = it.members["frameWidth"]
+                val height = it.members["frameHeight"]
+                val fps = it.members["framesPerSecond"]
+                val deviceLatency = it.members["totalPacketSendDelay"]
+                // fir pli nack are also interesting
+
+            }
+
+            statGroups["inbound-rtp:video"]?.firstOrNull()?.let {
+                val jitter = it.members["jitter"] as Double
+                subscriber._jitterInMs.value = (jitter * 1000).toInt()
+            }
+            statGroups["track:video"]?.forEach {
+                // trackIdentifier is a random UUID generated by the browser
+                // map to a participant?
+                val trackId = it.members["trackIdentifier"]
+                val participantId = trackToParticipant[trackId] ?: "unknown"
+                val participantVideo = displayingAt[participantId] ?: emptyMap()
+                val visibleAt = participantVideo[TrackType.TRACK_TYPE_VIDEO]
+                val freezeInSeconds = it.members["totalFreezesDuration"]
+                val frameWidth = it.members["frameWidth"] as? Long
+                val frameHeight = it.members["frameHeight"] as? Long
+                val received = it.members["framesReceived"] as? Long
+                val duration = it.members["totalFramesDuration"] as? Long
+
+                logger.i { "receiving video for ${participantId} at $frameWidth: ${it.members["frameWidth"]} and rendering it at ${visibleAt?.dimensions?.width} visible: ${visibleAt?.visible}" }
+
+            }
+        }
+
+        statGroups.forEach {
+            logger.i { "statgroup ${it.key}:${it.value}" }
+        }
+
+    }
+
+    fun updateLocalStats() {
+        val resolution = call.camera?.resolution?.value
+        val availableResolutions = call.camera?.availableResolutions?.value
+        val maxResolution = availableResolutions?.maxByOrNull { it.width * it.height }
+
+        val sfu = call.session?.sfuUrl
+
+        val sdk = "android"
+        // TODO: How do we get this? val version = Configuration.versionName
+        val osVersion = Build.VERSION.RELEASE ?: ""
+
+        val vendor = Build.MANUFACTURER ?: ""
+        val model = Build.MODEL ?: ""
+        val deviceModel = ("$vendor $model").trim()
+
+        val local = LocalStats(
+            resolution = resolution,
+            availableResolutions = availableResolutions,
+            maxResolution = maxResolution,
+            sfu = sfu ?: "",
+            os = osVersion,
+            sdkVersion = "0.1",
+            deviceModel = deviceModel,
+        )
+        _local.value = local
+
+
+    }
+
+}
